cleanup match logic 2x
